Attempt at making main.go a bit more managable
