+++ cmd/main.go
@@ -2,30 +2,20 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"flag"
 	"fmt"
 	"github.com/ardanlabs/conf/v3"
 	"github.com/ldebruijn/graphql-protect/internal/app/config"
-	"github.com/ldebruijn/graphql-protect/internal/business/aliases"
-	"github.com/ldebruijn/graphql-protect/internal/business/batch"
 	"github.com/ldebruijn/graphql-protect/internal/business/block_field_suggestions"
-	"github.com/ldebruijn/graphql-protect/internal/business/enforce_post"
-	"github.com/ldebruijn/graphql-protect/internal/business/gql"
-	"github.com/ldebruijn/graphql-protect/internal/business/max_depth"
 	"github.com/ldebruijn/graphql-protect/internal/business/persisted_operations"
+	"github.com/ldebruijn/graphql-protect/internal/business/protect"
 	"github.com/ldebruijn/graphql-protect/internal/business/schema"
-	"github.com/ldebruijn/graphql-protect/internal/business/tokens"
 	"github.com/ldebruijn/graphql-protect/internal/http/middleware"
 	"github.com/ldebruijn/graphql-protect/internal/http/proxy"
 	"github.com/ldebruijn/graphql-protect/internal/http/readiness"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/vektah/gqlparser/v2/ast"
-	"github.com/vektah/gqlparser/v2/gqlerror"
-	"github.com/vektah/gqlparser/v2/parser"
-	"github.com/vektah/gqlparser/v2/validator"
 	log2 "log"
 	"log/slog"
 	"net/http"
@@ -48,7 +38,6 @@ var (
 	},
 		[]string{"version", "go_version", "short_hash"},
 	)
-	errRedacted = errors.New("error(s) redacted")
 )
 
 func init() {
@@ -117,13 +106,19 @@ func run(log *slog.Logger, cfg *config.Config, shutdown chan os.Signal) error { 
 		return nil
 	}
 
+	protectHandler, err := protect.NewGraphQLProtect(log, cfg, po, schemaProvider, pxy)
+	if err != nil {
+		log.Error("Error initializing GraphQL Protect", "err", err)
+		return err
+	}
+
 	mux := http.NewServeMux()
 
-	mid := middlewareChain(log, cfg, po, schemaProvider)
+	mid := ProtectMiddlewareChain(log)
 
 	mux.Handle("/metrics", promhttp.Handler())
 	mux.Handle("/internal/healthz/readiness", readiness.NewReadinessHandler())
-	mux.Handle(cfg.Web.Path, mid(pxy))
+	mux.Handle(cfg.Web.Path, mid(protectHandler))
 
 	api := http.Server{
 		Addr:         cfg.Web.Host,
@@ -163,90 +158,13 @@ func run(log *slog.Logger, cfg *config.Config, shutdown chan os.Signal) error { 
 	return nil
 }
 
-func middlewareChain(log *slog.Logger, cfg *config.Config, po *persisted_operations.PersistedOperationsHandler, schema *schema.Provider) func(next http.Handler) http.Handler {
+func ProtectMiddlewareChain(log *slog.Logger) func(next http.Handler) http.Handler {
 	rec := middleware.Recover(log)
 	httpInstrumentation := middleware.RequestMetricMiddleware()
 
-	aliases.NewMaxAliasesRule(cfg.MaxAliases)
-	max_depth.NewMaxDepthRule(cfg.MaxDepth)
-	tks := tokens.MaxTokens(cfg.MaxTokens)
-	maxBatch, err := batch.NewMaxBatch(cfg.MaxBatch)
-	if err != nil {
-		log.Warn("Error initializing maximum batch protection", err)
-	}
-
-	vr := ValidationRules(schema, tks, maxBatch, cfg.ObfuscateValidationErrors)
-	disableMethod := enforce_post.EnforcePostMethod(cfg.EnforcePost)
-
 	fn := func(next http.Handler) http.Handler {
-		return rec(httpInstrumentation(disableMethod(po.Execute(vr(next)))))
+		return rec(httpInstrumentation(next))
 	}
 
 	return fn
-}
-
-func ValidationRules(schema *schema.Provider, tks *tokens.MaxTokensRule, batch *batch.MaxBatchRule, obfuscateErrors bool) func(next http.Handler) http.Handler { // nolint:funlen,cyclop
-	return func(next http.Handler) http.Handler {
-		fn := func(w http.ResponseWriter, r *http.Request) {
-			payload, err := gql.ParseRequestPayload(r)
-			if err != nil {
-				next.ServeHTTP(w, r)
-				return
-			}
-
-			var errs gqlerror.List
-
-			err = batch.Validate(payload)
-			if err != nil {
-				errs = append(errs, gqlerror.Wrap(err))
-			}
-
-			// only process the rest if no error yet
-			if err == nil {
-				for _, data := range payload {
-					operationSource := &ast.Source{
-						Input: data.Query,
-					}
-
-					err = tks.Validate(operationSource)
-					if err != nil {
-						errs = append(errs, gqlerror.Wrap(err))
-						continue // we could consider break-ing here. That would short-circuit on error, with the downside of not returning all potential errors
-					}
-
-					var query, err = parser.ParseQuery(operationSource)
-					if err != nil {
-						errs = append(errs, gqlerror.Wrap(err))
-						continue
-					}
-
-					errList := validator.Validate(schema.Get(), query)
-					if len(errList) > 0 {
-						errs = append(errs, errList...)
-						continue
-					}
-				}
-			}
-
-			if len(errs) > 0 {
-				if obfuscateErrors {
-					errs = gqlerror.List{gqlerror.Wrap(errRedacted)}
-				}
-
-				response := map[string]interface{}{
-					"data":   nil,
-					"errors": errs,
-				}
-
-				err = json.NewEncoder(w).Encode(response)
-				if err != nil {
-					log2.Println(err)
-				}
-				return
-			}
-
-			next.ServeHTTP(w, r)
-		}
-		return http.HandlerFunc(fn)
-	}
 }
