+++ cmd/main_test.go
@@ -331,7 +331,7 @@ input ImageInput {
 				expected := map[string]interface{}{
 					"errors": []map[string]interface{}{
 						{
-							"message": errRedacted.Error(),
+							"message": "error(s) redacted",
 						},
 					},
 				}
@@ -340,7 +340,7 @@ input ImageInput {
 				assert.NoError(t, err)
 				// perform string comparisons as map[string]interface seems incomparable
 				fmt.Println(string(actual))
-				assert.True(t, errorsContainsMessage(errRedacted.Error(), actual))
+				assert.True(t, errorsContainsMessage("error(s) redacted", actual))
 			},
 		},
 		{
